KL-149/style: separate city method to util
