+++ src/main/java/taco/klkl/domain/region/service/CityService.java
@@ -5,15 +5,12 @@ import java.util.Set;
 
 import org.springframework.stereotype.Service;
 
-import taco.klkl.domain.region.domain.City;
 import taco.klkl.domain.region.dto.response.CityResponse;
 import taco.klkl.domain.region.enums.CityType;
 
 @Service
 public interface CityService {
-	City getCityEntityById(final Long id);
 
 	List<CityResponse> getAllCitiesByCityTypes(final List<CityType> cityTypes);
 
-	boolean isCitiesMappedToSameCountry(final Set<Long> cityIds);
 }
